:truck: refactor services folder - matching api (#72)
